requirements.txt
absl-py==2.3.1
altair==6.0.0
astunparse==1.6.3
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.1
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
narwhals==2.14.0
numpy==2.2.6
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.15.1.9
nvidia-cufft-cu12==11.4.1.4
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-nccl-cu12==2.28.7
nvidia-nvjitlink-cu12==12.9.86
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.1
pyarrow==22.0.0
pydeck==0.9.1
pyparsing==3.2.5
PySocks==1.7.1
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
smmap==5.0.2
soupsieve==2.8
streamlit==1.52.1
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-addons==0.23.0
termcolor==3.2.0
threadpoolctl==3.6.0
toml==0.10.2
tqdm==4.67.1
typeguard==2.13.3
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==2.0.1
